Add input validation to validating webhook
The server should not emit a server error when it receives invalid
input, rather it should process and helpfully reject invalid input. This
change adds that client input validation and emits 4xx errors when it
receives invalid JSON from the request. This is helpful in case a
developer is testing a hand-crafted AdmissionReview object or in case
the kubernetes API server is in some other way misbehaving and sending
bad input.